Moved the Jump sound out of main and into the actual Jump Resolution. Extracted out the Param objects for less overhead
