+++ OpenGLTest/OpenGLTest.vcxproj.filters
@@ -235,6 +235,12 @@
     <ClInclude Include="MenuScene.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="SoundParams.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="TrackParams.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <Xml Include="..\UI\DefaultUI.xml">
