+++ OpenGLTest/main.cpp
@@ -72,17 +72,6 @@ void KeyCallback(GLFWwindow *window, int key, int scancode, int action, int mods
 			TypeParam<bool> param(false);
 			EventManager::notify(PLAYER_JUMP, &param, false);
             
-            SoundParams * JumpNoise = new SoundParams();
-
-            JumpNoise->sound = Jump;
-
-            JumpNoise->x = 0;
-            JumpNoise->y = 0;
-            JumpNoise->z = 0;
-
-            TypeParam<SoundParams*> *jumpSound = new TypeParam<SoundParams*>(JumpNoise);
-            EventManager::notify(PLAY_SOUND, jumpSound);
-            
 		}
 		break;
 	case GLFW_KEY_F:
